feat: rewrite as ESM and improve the TypeScript declaration file (#14)
And a new `keepLeadingDash` option has been added to the kebabCase function.
When set to `false` it will remove any leading dash from the kebab-cased string,
in case the original string starts with a leading uppercase letter.

Fixes #7, Closes #12.

BREAKING CHANGE: kebab-case is rewritten in ESM and is not published as CommonJS anymore. If you still need CJS you should stick to the v1 version.